Feature: build issue edit page
