+++ app/issues/_components/IssueForm.tsx
@@ -1,0 +1,77 @@
+"use client";
+
+import ErrorMessage from "@/app/components/ErrorMessage";
+import Spinner from "@/app/components/Spinner";
+import { createIssueSchema } from "@/app/validationSchema";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { Issue } from "@prisma/client";
+import { Button, Callout, TextField } from "@radix-ui/themes";
+import axios from "axios";
+import "easymde/dist/easymde.min.css";
+import dynamic from "next/dynamic";
+import { useRouter } from "next/navigation";
+import { useState } from "react";
+import { Controller, SubmitHandler, useForm } from "react-hook-form";
+import { z } from "zod";
+
+const SimpleMDE = dynamic(() => import("react-simplemde-editor"), {
+  ssr: false,
+});
+type IssueFormValues = z.infer<typeof createIssueSchema>;
+
+const IssueForm = ({ issue }: { issue?: Issue }) => {
+  const router = useRouter();
+  const {
+    register,
+    handleSubmit,
+    control,
+    formState: { errors },
+  } = useForm<IssueFormValues>({
+    resolver: zodResolver(createIssueSchema),
+  });
+  const [error, setError] = useState("");
+  const [isSubmitting, setSubmitting] = useState(false);
+
+  const onSubmit: SubmitHandler<IssueFormValues> = async (data) => {
+    try {
+      setSubmitting(true);
+      await axios.post("/api/issues", data);
+      router.push("/issues");
+    } catch (error) {
+      setSubmitting(false);
+      setError("An unexpected error occurred");
+    }
+  };
+
+  return (
+    <div className="max-w-xl">
+      {error && (
+        <Callout.Root className="mb-5">
+          <Callout.Text>{error}</Callout.Text>
+        </Callout.Root>
+      )}
+      <form className=" space-y-3" onSubmit={handleSubmit(onSubmit)}>
+        <TextField.Root
+          defaultValue={issue?.title}
+          placeholder="Enter a title"
+          {...register("title")}
+        />
+        <ErrorMessage>{errors.title?.message}</ErrorMessage>
+        <Controller
+          name="description"
+          defaultValue={issue?.description}
+          control={control}
+          render={({ field }) => (
+            <SimpleMDE placeholder="Description" {...field} />
+          )}
+        />
+        <ErrorMessage>{errors.description?.message}</ErrorMessage>
+        <Button disabled={isSubmitting} type="submit">
+          Submit new Issue {isSubmitting && <Spinner />}
+        </Button>
+      </form>
+    </div>
+  );
+};
+
+export default IssueForm;
